Refactor Firebase configuration to use modular SDK and remove deprecated code
